fix bug busqueda agenda por nombre

+++ BMS.WEB/pages/diary/RegisterDiary.aspx.cs
@@ -67,7 +67,7 @@ namespace BMS.WEB.pages.diary
                 Diary diary = new Diary()
                 {
                     Id_Functionary = Convert.ToDecimal(dicProperties["Id_Functionary"]),
-                    Id_Visitor = float.Parse(dicProperties["Id_Visitor"]),
+                    Id_Visitor = decimal.Parse(dicProperties["Id_Visitor"]),
                     Description = dicProperties["Description"],
                     DateDiary = Convert.ToDateTime(dicProperties["DateDiary"]).ToString(ConfigManager.FieldsTypeDateFormat),
                     HourDiary = Convert.ToDateTime(dicProperties["HourDiary"]).ToString(ConfigManager.FieldsTypeTimeFormat),
